Split native top bar view
